Refactor color option styles and adjust height; update dialog positioning; enhance delete dialog logic; streamline import local component; modify popup option dimensions; adjust content list padding; clean up plugin settings state; improve mouse event handling; refine dictionary error messaging.

+++ src/components/dialogs/deleteDialog/component.tsx
@@ -32,7 +32,11 @@ class DeleteDialog extends React.Component<
     this.props.handleFetchNotes();
   };
   handleComfirm = async () => {
-    if (this.props.mode === "shelf" && !this.state.isDeleteShelfBook) {
+    if (
+      this.props.mode === "shelf" &&
+      !this.state.isDeleteShelfBook &&
+      !this.state.isDisableTrashBin
+    ) {
       this.deleteBookFromShelf();
     } else if (this.props.mode === "trash") {
       await this.deleteAllBookInTrash();
